% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_average_model}
\alias{add_average_model}
\title{Do very basic model averaging}
\usage{
add_average_model(pred_results)
}
\arguments{
\item{y}{}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
