% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/utilities.R
\name{analysis_one_country}
\alias{analysis_one_country}
\title{Do entire analysis for one country}
\usage{
analysis_one_country(y, country = "country", plot = F)

analysis_one_country(y, country = "country", plot = F)
}
\arguments{
\item{y}{}

\item{y}{}
}
\value{
Missing

Missing
}
\description{
Missing. 
Also Todo: handling for only one item

Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
NULL
}
