% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{bsts_local_trend}
\alias{bsts_local_trend}
\title{BSTS with Stan and local trend}
\usage{
bsts_local_trend(y, num_pred = 10, vb = FALSE, length_local_trend,
  prior_var_phi = 0.1, ...)
}
\arguments{
\item{y}{Numeric vector of output values, i.e. vector of a series of R_t values}

\item{num_pred}{number of time points to forecast}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
BSTS with Stan and local trend
}
