% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bsts_wrapper}
\alias{bsts_wrapper}
\title{Wrapper around the functions from the bsts package}
\usage{
bsts_wrapper(y, model, num_pred = 7, n_iter = 2000)
}
\arguments{
\item{y}{}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
