% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_bsts_models}
\alias{compare_bsts_models}
\title{Wrapper to select the best bsts model from the package}
\usage{
compare_bsts_models(y)
}
\arguments{
\item{y}{}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
