% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_forecasts}
\alias{compare_forecasts}
\title{Wrapper to compare forecasts}
\usage{
compare_forecasts(pred_results)
}
\arguments{
\item{y}{}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
