% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{do_all_fits}
\alias{do_all_fits}
\title{Wrapper to do all the fits that I have}
\usage{
do_all_fits(y, include_stan = FALSE, models = NULL)
}
\arguments{
\item{y}{}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
