% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract Predictive Samples From Stanobject}
\usage{
extract_samples(stanfitobject, predictive = TRUE)
}
\arguments{
\item{stanfitobject}{a stanfitobject}

\item{predictive}{TRUE indicates that you want to extract
predictive values, FALSE returns fitted values}
}
\value{
a matrix of dimension nxN with predictive samples.
n is the number of predicted values, N is the number of
MCMC samples for each value to be predicted.
}
\description{
Extract preditive samples and bring them in a format that 
can be used for scoring. Function needs to be updated to 
deal with more than one predictive parameter?
}
\examples{
NULL
}
