% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fit_iteratively}
\alias{fit_iteratively}
\title{Iteratively fit a model to the data}
\usage{
fit_iteratively(incidences, n_pred = 14, interval = NULL,
  start_period = 15, max_n_past_obs = Inf, model = NULL, ...)
}
\arguments{
\item{incidences}{Vector of length n with the past incidences used to 
fit the model and make predictions.}

\item{n_pred}{prediction horizon, i.e. number of days to predict into the 
future}

\item{interval}{interval between predictions. Maybe delete?}

\item{max_n_past_obs}{maximum number of past observations to take into 
account}
}
\value{
a data.frame with the data
}
\description{

}
\examples{
NULL
}
