% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fit_stan_model}
\alias{fit_stan_model}
\title{Fit Stan model}
\usage{
fit_stan_model(y, model, n_pred, vb, iter = 4000, length_local_trend,
  ...)
}
\arguments{
\item{y}{Vector of length n with the true values
fit the model and make predictions.}

\item{model}{Missing}
}
\value{
Missing
}
\description{
Wrapper around different lower level fit functions
}
\examples{
NULL
}
