% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/utilities.R
\name{forecast_one_country}
\alias{forecast_one_country}
\title{Wrapper to do all the fits that I have}
\usage{
forecast_one_country(y, include_stan = FALSE, models = NULL)

forecast_one_country(y, include_stan = FALSE, models = NULL)
}
\arguments{
\item{y}{}

\item{y}{}
}
\value{
Missing

Missing
}
\description{
Missing. 
Also Todo: handling for only one item

Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
NULL
}
