% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{forecast_table}
\alias{forecast_table}
\title{Extract summary data.frame from forecasts}
\usage{
forecast_table(pred_result, country = "country")
}
\arguments{
\item{y}{}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
