% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/utilities.R
\name{full_analysis}
\alias{full_analysis}
\title{Do the full analysis}
\usage{
full_analysis(timeseries, countries)

full_analysis(timeseries, countries)
}
\arguments{
\item{y}{}

\item{y}{}
}
\value{
Missing

Missing
}
\description{
Missing. 
Also Todo: handling for only one item

Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
NULL
}
