% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_data}
\alias{get_data}
\title{Load Wuhan Data}
\usage{
get_data()
}
\value{
a data.frame with the data
}
\description{
wrapper around the nCov2019 pacakge to load the Wuhan data and calculates
daily cases
}
\examples{
NULL
}
