% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_forecast_compare}
\alias{plot_forecast_compare}
\title{Wrapper to make a visual plot to compare the outputs of do_all_fits}
\usage{
plot_forecast_compare(pred_results)
}
\arguments{
\item{y}{}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
