% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_pred_vs_true}
\alias{plot_pred_vs_true}
\title{Plot predictive samples vs. true values}
\usage{
plot_pred_vs_true(y_true, y_pred_samples, forecast_run = NULL)
}
\arguments{
\item{y_true}{Vector of length n with the true values
fit the model and make predictions.}

\item{y_pred_samples}{predictive samples}

\item{forecast_run}{vector indicating which forecast run an 
esimtate belongs to}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: add an option for point estimates
get the forecast_run thing tidy
}
\examples{
NULL
}
