% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_prior_vs_posterior}
\alias{plot_prior_vs_posterior}
\title{Plot prior vs posterior distribution of parameters}
\usage{
plot_prior_vs_posterior(stanfitobjects)
}
\arguments{
\item{stanfitobjects}{list of stanfit objects}
}
\value{
Missing
}
\description{
Missing. 
Also Todo: handling for only one item
}
\examples{
NULL
}
