% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{linear_regression}
\alias{linear_regression}
\title{Bayesian linear regression with Stan}
\usage{
linear_regression(y, x = NULL, num_pred = 10, ...)
}
\arguments{
\item{y}{Numberic vector of output values.}

\item{x}{Numeric vector of input values.}

\item{num_pred}{number of time points to forecast}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian linear regression with Stan
}
